requirements.txt
# Core dependencies
hyperliquid-python-sdk>=0.20.0
openai>=1.0.0  # For Deepseek API (OpenAI compatible)

# Data processing
pandas>=2.0.0
numpy>=1.24.0
ta-lib>=0.4.0  # Technical analysis library

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Task scheduling
apscheduler>=3.10.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Logging
colorlog>=6.7.0

# Data visualization (for backtesting)
matplotlib>=3.7.0
plotly>=5.14.0

# Database (optional)
sqlalchemy>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
